File path issue ( / for \ ) corrected for linux and mac
The code was missing a check for the use of backslashes/forwardslashes
in file paths on linux/max/windows. There was already a check but it
still threw an error. This is fixed now. Ran all regression tests on
both Windows 10 (MATLAB 2016b) and Linux Debian (MATLAB 2017a) and they
all passed.