Stress on obs points matrix calc, fast eig funcs and cleaned code
Added a function to calculate the influence matrix of a surface and its
effect on observation points, this is not as fast as the ususal calc
unless you are using loops and presaved influence mats. Then there is a
significant speed up.
I have added some faster eig functions for 2*2 and 3*3 matricies. These
now dont use the eig func specifically but faster seperate funcs (matrix
decomposition combined with cross products).
I have also cleaned up the influence builder functions so these are
easier to read and are far shorter (rely on just a few external function
calls).