Complete code overhaul
Every function now has a description, inputs and outputs with example
cases (test cases and how to use).
New functions for plotting, also all figures now have white backgrounds
(not MATLAB default).
Changed how users add file paths, these are no longer duplicated in
every script but now this is in a script the user must run first.
Cleaned and reduced the size of all the regression tests. Analytical
solutions are now all functions.
Made all the scripts so these are within MATLABs default confines (fit
small screens without having to scroll to the right).
Changed the inhomogeneous elastic equation so its correct for
displacement.
Changed the friction solver in 2D and 3D so the influence matrices are
scaled so the solver works correctly (scaled by element size and the
stiffness of the elastic).
In the stress influence builder, this now uses structures to contain the
influence matrices so the function calls are not as large. This used to
be particularly apparent for the inhomogeneous influence builder.
The 2D and 3D codes are internally consistent. If you learn to use one
now it will be simple to use the other.
Updated the readme file so it describes common variable names in the
code and what these are. Less common variables will be inside functions
which now have their own input and out descriptions.
Reduced the size of the regression test pictures and changed the gif so
now the webpage should load faster and be more responsive to users
scrolling.
Changed the name of the software from “BEM_DDM_MATLAB” to
“CutAndDisplace”.