Quick fix to Mehdis artefact free functions
Artefacts still exist in certain edge cases when dislocation leg is close to vertical and a observation point lies under the dislocation line between the leg and vertical. This only occurs when this leg->vertical angle is very small. The following fix stops this happening if the leg is ~ 10000th of a degree away from vertical. When the angle is larger it appears the 'cot' function doesnt blow up so is 'safe' but there may still be some error involved, a more rigorous fix from Mehdi is to come....